final code
code with comments
